x64 compatibility fixes
git-svn-id: https://svn.code.sf.net/p/deinterlace/code/trunk/DScaler5@4654 6e12ab15-cf2f-4ffe-bf36-45892a6a5b89

+++ Filters/Source/TSReader/TSReader.cpp
@@ -545,7 +545,7 @@ void CTSReader::ProcessingThread(void* pParam)
         hr = pThis->PushFile(hFile);
         CloseHandle(hFile);
 
-        long namelength(wcslen(pThis->m_FileName));
+        size_t namelength(wcslen(pThis->m_FileName));
 
         if(namelength >= 4)
         {
